+++ QueryBuilder.php
@@ -185,14 +185,14 @@ class QueryBuilder extends \yii\base\Object
             'not like' => 'buildLikeCondition',
             'or like' => 'buildLikeCondition',
             'or not like' => 'buildLikeCondition',
-			'lt' => 'buildUnboundRangeCondition',
-			'<' => 'buildUnboundRangeCondition',
-			'lte' => 'buildUnboundRangeCondition',
-			'<=' => 'buildUnboundRangeCondition',
-			'gt' => 'buildUnboundRangeCondition',
-			'>' => 'buildUnboundRangeCondition',
-			'gte' => 'buildUnboundRangeCondition',
-			'>=' => 'buildUnboundRangeCondition',
+            'lt' => 'buildHalfBoundedRangeCondition',
+            '<' => 'buildHalfBoundedRangeCondition',
+            'lte' => 'buildHalfBoundedRangeCondition',
+            '<=' => 'buildHalfBoundedRangeCondition',
+            'gt' => 'buildHalfBoundedRangeCondition',
+            '>' => 'buildHalfBoundedRangeCondition',
+            'gte' => 'buildHalfBoundedRangeCondition',
+            '>=' => 'buildHalfBoundedRangeCondition',
         ];
 
         if (empty($condition)) {
@@ -348,38 +348,49 @@ class QueryBuilder extends \yii\base\Object
         return $filter;
     }
 
-    private function buildUnboundRangeCondition($operator, $operands)
+    /**
+     * Builds a half-bounded range condition
+     * (for "gt", ">", "gte", ">=", "lt", "<", "lte", "<=" operators)
+     * @param string $operator
+     * @param array $operands
+     * @return array Filter expression
+     */
+    private function buildHalfBoundedRangeCondition($operator, $operands)
     {
-    	if (!isset($operands[0], $operands[1])) {
-    		throw new InvalidParamException("Operator '$operator' requires two operands.");
-    	}
+        if (!isset($operands[0], $operands[1])) {
+            throw new InvalidParamException("Operator '$operator' requires two operands.");
+        }
 
-    	list($column, $value) = $operands;
-    	if ($column == '_id') {
-    		$column = '_uid';
-    	}
+        list($column, $value) = $operands;
+        if ($column == '_id') {
+            $column = '_uid';
+        }
 
-    	$range_operator = 'gte';
+        $range_operator = null;
 
-    	if (in_array($operator, ['gte', '>='])) {
-    		$range_operator = 'gte';
-    	} else if (in_array($operator, ['lte', '<='])) {
-    		$range_operator = 'lte';
-    	} else if (in_array($operator, ['gt', '>'])) {
-    		$range_operator = 'gt';
-    	} else if (in_array($operator, ['lt', '<'])) {
-    		$range_operator = 'lt';
-    	}
+        if (in_array($operator, ['gte', '>='])) {
+            $range_operator = 'gte';
+        } else if (in_array($operator, ['lte', '<='])) {
+            $range_operator = 'lte';
+        } else if (in_array($operator, ['gt', '>'])) {
+            $range_operator = 'gt';
+        } else if (in_array($operator, ['lt', '<'])) {
+            $range_operator = 'lt';
+        }
 
-    	$filter = [
-    		'range' => [
-    			"$column" => [
-    				"$range_operator" => $value
-    			]
-    		]
-    	];
+        if ($range_operator === null) {
+            throw new InvalidParamException("Operator '$operator' is not implemented.");
+        }
 
-    	return $filter;
+        $filter = [
+            'range' => [
+                $column => [
+                    $range_operator => $value
+                ]
+            ]
+        ];
+
+        return $filter;
     }
 
     protected function buildCompositeInCondition($operator, $columns, $values)